add(svg-maps/common): shared types for @svg-maps
This extracts common interfaces used accross @svg-maps packages into
definition only repo file, that could be shared accross all packages
implementing @svg-maps/* on DT

/cc @nickglazer for your review

Thanks!
