+++ types/svg-maps__usa/svg-maps__usa-tests.ts
@@ -3,19 +3,23 @@ import USA, { Map } from '@svg-maps/usa';
 const map: Map = USA;
 
 const SVGMapContainer: Map = {
-  viewBox: 'viewBox',
-  locations: [{
-    path: 'path',
-    id: 'id',
-    name: 'name'
-  }],
-  label: 'label'
+    viewBox: 'viewBox',
+    locations: [
+        {
+            path: 'path',
+            id: 'id',
+            name: 'name',
+        },
+    ],
+    label: 'label',
 };
 
 const SVGMapContainerNoOptional: Map = {
-  viewBox: 'vieweBox',
-  locations: [{
-  path: 'path',
-  id: 'id',
-  }],
+    viewBox: 'viewBox',
+    locations: [
+        {
+            path: 'path',
+            id: 'id',
+        },
+    ],
 };
